lib: nrf_modem: remove `lte_net_if.h` header
The `lte_net_if` component (formerly `lte_connectivity`) does not need
its own header, since it does not add any API for the use. Rather,
it provides support for Zephyr's net_if API.

Remove the `lte_net_if.h` file, and move the macros for the network
interface options to `nrf_modem_lib.h`.
